+++ Runtime/AnimationRig/Constraints/TwistChainConstraint.cs
@@ -42,7 +42,7 @@ namespace UnityEngine.Animations.Rigging
     /// TwistChain constraint
     /// </summary>
     [DisallowMultipleComponent, AddComponentMenu("Animation Rigging/Twist Chain Constraint")]
-    [HelpURL("https://docs.unity3d.com/Packages/com.unity.animation.rigging@1.1/manual/constraints/TwistChainConstraint.html")]
+    [HelpURL("https://docs.unity3d.com/Packages/com.unity.animation.rigging@1.2/manual/constraints/TwistChainConstraint.html")]
     public class TwistChainConstraint : RigConstraint<
         TwistChainConstraintJob,
         TwistChainConstraintData,